+++ lib/pharos/root_command.rb
@@ -4,7 +4,6 @@ require_relative 'up_command'
 require_relative 'reset_command'
 require_relative 'version_command'
 require_relative 'kubeconfig_command'
-require_relative 'license_command'
 
 module Pharos
   class RootCommand < Pharos::Command
@@ -13,7 +12,6 @@ module Pharos
     subcommand ["build", "up"], "initialize/upgrade cluster", UpCommand
     subcommand "kubeconfig", "fetch admin kubeconfig file", KubeconfigCommand
     subcommand "reset", "reset cluster", ResetCommand
-    subcommand "license", "license subcommands", LicenseCommand
     subcommand ["version"], "show version information", VersionCommand
 
     def self.run
@@ -24,3 +22,5 @@ module Pharos
     end
   end
 end
+
+Dir.glob(File.join(__dir__, '..', '..', 'non-oss', 'commands', '*.rb')).each { |non_oss_command| require non_oss_command }
