+++ non-oss/commands/license_assign_command.rb
