+++ non-oss/spec/commands/license_assign_command_spec.rb
